Clean up Auth Components
I changed a lot. I made a new container and separated out some JSX and functions to keep separation of concerns clean and clear. I then made sure the props are passing down from App.js.
