+++ neighborhood-library/src/components/Account/Account.js
@@ -1,29 +1,29 @@
-import React from "react";
+import React, { useState, useEffect } from "react";
 import { Redirect } from "react-router-dom";
 import { StripeProvider, Elements } from "react-stripe-elements";
 import Stripe from "./Stripe/Stripe";
 import UserInfo from "./UserInfo";
 
-const Account = () => {
+const Account = props => {
   return (
     <div>
-      <h2>{this.props.name}</h2>
+      <h2>{props.name}</h2>
       <UserInfo
-        userId={this.props.userId}
-        name={this.props.name}
-        firstName={this.props.firstName}
-        lastName={this.props.lastName}
-        email={this.props.email}
-        address={this.props.address}
-        city={this.props.city}
-        props={this.props.state}
-        zipcode={this.props.zipcode}
-        img={this.props.img}
+        userId={props.userId}
+        name={props.name}
+        firstName={props.firstName}
+        lastName={props.lastName}
+        email={props.email}
+        address={props.address}
+        city={props.city}
+        props={props.state}
+        zipcode={props.zipcode}
+        img={props.img}
       />
 
       <StripeProvider apiKey="pk_test_j6wi0FWmtWCqFPwU3oCHJA2800c8YshuOy">
         <Elements>
-          <Stripe stripe_user_id={this.props.stripe_user_id} />
+          <Stripe stripe_user_id={props.stripe_user_id} />
         </Elements>
       </StripeProvider>
     </div>
