+++ neighborhood-library/src/components/Auth/Login.js
@@ -2,8 +2,8 @@ import React, { Component } from "react";
 import AuthPostData from "./AuthPostData";
 
 class Login extends Component {
-  constructor() {
-    super();
+  constructor(props) {
+    super(props);
     this.state = {
       password: "",
       email: ""
@@ -49,6 +49,16 @@ class Login extends Component {
           <button type="submit">Login</button>
         </form>
         <a href="/change-password">Forgot password?</a>
+        <div>
+          <p>Need an account?</p>
+          <button
+            href="#"
+            className="manual-auth-nav"
+            onClick={this.loginVsRegisterToggler}
+          >
+            Register
+          </button>
+        </div>
       </div>
     );
   }
